Wrap the encrypt hook to simplify usage and control cursor position
This changes the way mkinitcpio-welcome is used. By replacing the
encrypt hook, the old poscursor hook is unnecessary as everything can be
contained in one 'welcome' hook. The encrypt hook's build and run_hook
functions are executed directly by this hook. This makes it easier to
control the position of the cursor and any messages printed by the
decryption step.